backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
python-telegram-bot
httpx
pydantic
python-dotenv
firecrawl-py
trafilatura
playwright
